refactor(graph-db): group daemon/*Lifecycle.ts under daemon/lifecycle/
The daemon dir had 16 immediate children, exceeding the 15-child
fanout limit enforced by packages/systems/codebase-shape.test.ts and
introduced by 4c4020e6 ("fix(graph-db): stop leaking vt-graphd daemons
across Electron crashes"). Move the 4 daemon*Lifecycle.ts files into a
new lifecycle/ subdir, leaving 13 children in daemon/ (good headroom).

The 7 imports that previously crossed `../../` boundaries are now
routed through @vt/graph-db-server/{graph,state}/* subpath aliases —
all of which already exist in the package's exports and are already
used cross-package, so no new public API is added. This keeps
bannedRelativeImports at 0 in the relative-imports report.

Files moved:
  daemon/daemon{Graph,Lock,Port,Watcher}Lifecycle.ts
    → daemon/lifecycle/daemon{Graph,Lock,Port,Watcher}Lifecycle.ts

Verified:
  - codebase-shape test passes
  - daemonWatcherLifecycle test passes
  - package-boundaries test passes (no new boundary violations)
  - All 48 graph-db-server daemon unit tests pass
  - measure-relative-imports: 0 banned
